+++ src/templates/TemplateModelsProvider.tsx
@@ -2,10 +2,11 @@ import { Dictionary } from 'lodash';
 import React from 'react';
 import { TemplatesMap } from 'types/Annotation';
 import { FeatureModel } from 'types/AnnotationModel';
+import { AvailableCharacteristic } from 'feature/data/getAvailableColumns';
 
 type ContextType = {
   templateModels: TemplatesMap;
-  availableColumns: Dictionary<Set<string>>;
+  availableColumns: Dictionary<AvailableCharacteristic>;
   editedWindow?: string;
   editedTemplateId?: number;
   setEditedWindow: (windowUid?: string) => void;